requirements.txt
# Keras 3 + PyTorch backend (works on Python 3.14+). For TensorFlow backend, install
# tensorflow and set HDFD_KERAS_BACKEND=tensorflow before importing keras.
keras>=3.5.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless>=4.9.0
xgboost>=2.0.0
pandas>=2.0
numpy>=1.26
scikit-learn>=1.5.0
pillow>=10.0
matplotlib>=3.8
tqdm>=4.66
flask>=3.0
flask-cors>=4.0
python-multipart>=0.0.9
joblib>=1.3
duckduckgo-search>=6.0
